backend/requirements.txt
# ============================================================
# TEAR - Token-Efficient Adaptive Retrieval for Scalable RAG
# Requirements File
# ============================================================

# Core ML / NLP
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0        # 4-bit quantization for Llama-3 70B
sentence-transformers>=2.7.0

# Vector Stores
faiss-cpu>=1.8.0            # use faiss-gpu if CUDA available
chromadb>=0.5.0

# Sparse Retrieval
rank_bm25>=0.2.2

# NLP Utilities
spacy>=3.7.0
# After install: python -m spacy download en_core_web_sm

# Dataset Loading
datasets>=2.19.0
huggingface_hub>=0.23.0

# API Backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# Evaluation Metrics
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
numpy>=1.26.0
scipy>=1.13.0
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.2

# Dev / Testing
pytest>=8.2.0
httpx>=0.27.0
